SRC/services/notification/requirements.txt
Django==5.0.10
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
django-filter==24.2
psycopg[binary]==3.2.1
celery==5.4.0
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.2.2
pytest-django==4.8.0
